Create homework

+++ src/main/java/mod_11/Task_4/RandomAlg.java
@@ -9,13 +9,13 @@ public class RandomAlg {
 		this.a = 25214903917l;
 		this.m = 2 ^ 48;
 	}
-	public RandomAlg withSeed (int c){
-		this.c = 11;
+	public RandomAlg withSeed (long c){
+		this.c = 11l;
 		x = new AtomicLong(c);
 		return this;
 	}
 
-	public  int next(){
-		return (int) ((a * x.getAndIncrement() + c) % m);
+	public  long next(){
+		return (a * x.getAndIncrement() + c) % m;
 	}
 }
